pyproject.toml
[project]
name = "confluence-mcp"
version = "0.1.0"
description = "Atlassian Confluence Cloud MCP server based on the REST API v2"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.26.0",
    "python-dotenv>=1.2.2",
]

[project.scripts]
confluence-mcp = "confluence_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/confluence_mcp"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-httpx>=0.36.0",
]
